+++ models/contacts.json
@@ -72,9 +72,9 @@
     "phone": "(380) 999999-999999"
   },
   {
-    "id": "GEbGApmsqB74lHMAHd6Sr",
-    "name": "Vasya",
-    "email": "vasya@mail.com",
-    "phone": "(380) 999999-999999"
+    "id": "_i2V941bkQmU_ZoNKofp6",
+    "name": "Timofey",
+    "email": "tima@mail.com",
+    "phone": "(380) 777777777777777777"
   }
 ]
