+++ server.log
@@ -64,3 +64,16 @@
 10-09-2023_03:28:16 DELETE /api/ptqHsa30HfRjNWhXMDVjY
 10-09-2023_03:28:47 DELETE /api/contacts/ptqHsa30HfRjNWhXMDVjY
 10-09-2023_03:28:57 DELETE /api/contacts/ptqHsa30HfRjNWhXMDVj
+10-09-2023_04:56:10 GET /api/contacts/
+10-09-2023_04:58:10 GET /api/contacts/
+10-09-2023_04:58:11 GET /api/contacts/
+10-09-2023_05:00:25 GET /api/contacts/LnUlH5O923i-NRDaKfxe_
+10-09-2023_05:03:12 POST /api/contacts/
+10-09-2023_05:05:14 GET /api/contacts/
+10-09-2023_05:05:39 GET /api/contacts/GEbGApmsqB74lHMAHd6Sr
+10-09-2023_05:06:32 DELETE /api/contacts/GEbGApmsqB74lHMAHd6Sr
+10-09-2023_05:08:28 GET /api/contacts/GEbGApmsqB74lHMAHd6Sr
+10-09-2023_05:08:32 GET /api/contacts/
+10-09-2023_05:09:31 PUT /api/contacts/_i2V941bkQmU_ZoNKofp6
+10-09-2023_05:09:43 GET /api/contacts/_i2V941bkQmU_ZoNKofp6
+10-09-2023_05:09:52 GET /api/contacts/
